+++ packages/client/src/components/Game/utils/rectangularCollision.ts
@@ -1,8 +1,9 @@
-import { Boundary, Food, Sprite } from '../classes'
+import { GameObject } from '../classes'
 
+// Функция для определения пересечения двух прямоугольников
 export function rectangularCollision(
-  rectangle1: Sprite | Boundary | Food,
-  rectangle2: Sprite | Boundary | Food
+  rectangle1: GameObject,
+  rectangle2: GameObject
 ): boolean {
   return (
     rectangle1.position.x + rectangle1.width >= rectangle2.position.x &&
